update build/test scripts for windows and unix

+++ build.sh
@@ -1,14 +1,27 @@
 # run from top level (sprint3)
 
+for arg in "$@"; do
+    if ["$arg" = "-clean" ]; then
+        rm -rf build
+        break
+    fi
+done
 
-rm -rf build
-mkdir build
+mkdir -p build
 cd build
 
+cmake -DCMAKE_BUILD_TYPE=Debug ..
 for arg in "$@"; do
     if [ "$arg" = "-min" ]; then
         cmake -DMAKE_MIN:BOOL=ON ..
-        break
+        continue
+    fi
+    if ["$arg" = "-test" ]; then
+        cmake -DMAKE_TEST:BOOL=ON ..
+        continue
+    fi
+    if ["$arg" = "-release" ]; then
+        cmake -DCMAKE_BUILD_TYPE=Release ..
     fi
 done
